First clean compile of Template Packaging
Implemented the concept of a “Package” of “Collections” as a
collection->package hash in the AbstractPersistence class. Added new
configuration parameter for the “Default Package” to be used when
adding new templates to a system. Error checking to make sure all
members of a collection are in the same package.